refactor(FileScoreRepository): update to use options
Refactor the `FileScoreRepository` to accept a `FileScoreRepositoryOptions` object, enhancing flexibility by allowing configuration of the path and separator. The separator is now a class member with a default value of "#&#". The `Append` and `TryParseLine` methods have been updated accordingly.

Additionally, modify the `BullsCowsText` class to accept a `codeLength` parameter in the `ToMarkers` method. Update the `IOutputFormatter` interface and `OutputFormatter` class to reflect this change. Finally, adjust the `Program` class to register the `IScoreRepository` using the new options class.

BREAKING CHANGE: The `FileScoreRepository` constructor signature has changed, requiring a `FileScoreRepositoryOptions` object instead of a string path.